# clarityhubbackend/docs/openapi_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Clarity Store API",
    "version": "1.0.0",
    "description": "Comprehensive E-commerce API Documentation for Clarity Store",
    "contact": {
      "name": "Clarity Support",
      "email": "support@claritystore.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Access token stored in HttpOnly cookie"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/users/customers": {
      "get": {
        "summary": "Get all customers (Admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of customers."
          }
        }
      }
    },
    "/users/customers/{id}": {
      "get": {
        "summary": "Get customer details (Admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Customer details."
          }
        }
      }
    },
    "/users/customers/{id}/status": {
      "patch": {
        "summary": "Update customer status (Admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "blocked"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated."
          }
        }
      }
    },
    "/users/profile/overview": {
      "get": {
        "summary": "Get admin profile (Admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Admin profile."
          }
        }
      }
    },
    "/users/profile/update": {
      "put": {
        "summary": "Update admin profile (Admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated."
          }
        }
      }
    },
    "/users/profile/audit-logs": {
      "get": {
        "summary": "Get audit logs (Admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Audit logs."
          }
        }
      }
    },
    "/users/profile/notifications": {
      "put": {
        "summary": "Update notification settings (Admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "emailNotifications": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated."
          }
        }
      }
    },
    "/users/profile/2fa": {
      "put": {
        "summary": "Toggle 2FA (Admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enable": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA updated."
          }
        }
      }
    },
    "/users/profile/api-keys": {
      "post": {
        "summary": "Generate API Key (Admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API Key generated."
          }
        }
      }
    },
    "/users/profile/api-keys/{id}": {
      "delete": {
        "summary": "Revoke API Key (Admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "API Key revoked."
          }
        }
      }
    },
    "/users/profile/sessions": {
      "get": {
        "summary": "Get active sessions (Admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Active sessions."
          }
        }
      }
    },
    "/users/profile/sessions/{id}": {
      "delete": {
        "summary": "Revoke session (Admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Session revoked."
          }
        }
      }
    }
  },
...
}